- 테이블 변경에 따른 엔티티 변경

+++ src/main/java/today/creame/web/home/entrypoint/io/InfluencesWithReviewResponse.java
@@ -20,7 +20,7 @@ public class InfluencesWithReviewResponse {
     private int rateCount;
     private int reviewCount;
     private int qnaCount;
-    private int item;
+    private Long item;
     private String m2NetCounselorId;
     // TODO: 제거하기
     private int coinPrice;
